papers/gniewkowskiSec2vecAnomalyDetection2023/requirements.txt
aiohttp==3.7.4.post0
aiohttp-retry==2.4.5
allennlp==2.2.0
allennlp-models==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
atpublic==2.1.3
attrs==20.3.0
backcall==0.2.0
bertviz==1.0.0
bleach==3.3.0
blis==0.7.4
boto3==1.17.42
botocore==1.20.42
cached-property==1.5.2
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
configparser==5.0.2
conllu==4.4
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
defusedxml==0.7.1
dictdiffer==0.8.1
diskcache==5.2.1
distro==1.5.0
docker-pycreds==0.4.0
dpath==2.0.5
dulwich==0.20.25
dvc==2.7.4
entrypoints==0.3
filelock==3.0.12
flatten-dict==0.4.2
flufl.lock==3.2
fsspec==2021.10.0
ftfy==5.9
funcy==1.15
future==0.18.2
gensim==4.1.0
gitdb==4.0.7
GitPython==3.1.14
grandalf==0.6
h5py==3.2.1
idna==2.10
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonnet==0.17.0
jsonpath-ng==1.5.2
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lmdb==1.1.1
mailchecker==4.0.6
MarkupSafe==1.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
more-itertools==8.7.0
multidict==5.1.0
murmurhash==1.0.5
nanotime==0.5.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
nltk==3.5
notebook==6.3.0
numpy==1.20.2
orjson==3.5.2
overrides==3.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pathlib2==2.3.5
pathspec==0.8.1
pathtools==0.1.2
pathy==0.4.0
pexpect==4.8.0
phonenumbers==8.12.20
pickleshare==0.7.5
Pillow==8.1.2
plotly==4.14.3
pluggy==0.13.1
ply==3.11
preshed==3.0.5
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
py-rouge==1.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pydot==1.4.2
pygit2==1.5.0
Pygments==2.8.1
pygtrie==2.3.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-benedict==0.23.2
python-dateutil==2.8.1
python-fsutil==0.4.0
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.3.17
requests==2.25.1
retrying==1.3.3
rich==10.0.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.2
s3transfer==0.3.6
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.2
Send2Trash==1.5.0
sentencepiece==0.1.95
sentry-sdk==1.0.0
shortuuid==1.0.1
shtab==1.3.5
simplejson==3.17.2
six==1.15.0
smart-open==3.0.0
smmap==4.0.0
spacy==3.0.5
spacy-legacy==3.0.1
srsly==2.4.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboardX==2.1
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
thinc==8.0.2
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
transformers==4.4.2
typer==0.3.2
typing-extensions==3.10.0.2
ujson==4.0.2
urllib3==1.26.4
voluptuous==0.12.1
wandb==0.10.24
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
word2number==1.1
xmltodict==0.12.0
yarl==1.6.3
zc.lockfile==2.0
shap==0.40.0
